rag-pdf-parse/requirements.txt
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
joblib==1.4.2
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
pdfminer.six==20231228
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
